requirements.txt
abstract_singleton==1.0.1
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.4
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blis==0.7.9
build==0.10.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
coverage==7.2.5
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
docutils==0.20
duckduckgo-search==2.9.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flake8==6.0.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
identify==2.5.24
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.9.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==2.0.5
mkdocs==1.4.3
more-itertools==9.1.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
newsapi-python==0.2.7
nltk==3.8.1
nodeenv==1.7.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
openpyxl==3.1.2
orjson==3.8.10
outcome==1.2.0
packaging==23.0
panda==0.3.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==3.5.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.3.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.17.4
pymdown-extensions==9.11
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
pytest-xdist==3.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-lorem==1.3.0
python-nmap==0.7.1
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.0.0
readability-lxml==0.8.1
readme-renderer==37.3
redis==4.5.4
regex==2023.5.4
requests==2.29.0
requests-mock==1.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
selenium==4.1.4
Send2Trash==1.8.0
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
summarizer==0.0.7
terminado==0.17.1
thinc==8.1.10
tiktoken==0.3.3
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
twine==4.0.2
typer==0.7.0
typing_extensions==4.5.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
validators==0.20.0
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
wrapt==1.15.0
wsproto==1.2.0
XlsxWriter==3.0.9
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
